[OPIK-2230] move to pages-shared
