+++ apps/opik-frontend/src/components/pages-shared/onboarding/IntegrationExplorer/IntegrationCopyApiKey.tsx
