+++ apps/opik-frontend/src/components/pages-shared/onboarding/IntegrationExplorer/IntegrationGetHelp.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Button } from "@/components/ui/button";
 import { HelpCircle } from "lucide-react";
-import HelpGuideDialog from "@/components/shared/IntegrationExplorer/components/HelpGuideDialog";
-import { useIntegrationExplorer } from "@/components/shared/IntegrationExplorer/IntegrationExplorerContext";
+import HelpGuideDialog from "@/components/pages-shared/onboarding/IntegrationExplorer/components/HelpGuideDialog";
+import { useIntegrationExplorer } from "@/components/pages-shared/onboarding/IntegrationExplorer/IntegrationExplorerContext";
 
 type IntegrationGetHelpProps = {
   className?: string;
