+++ apps/opik-frontend/src/components/pages-shared/onboarding/IntegrationExplorer/components/HelpGuideDialog.tsx
