+++ apps/opik-frontend/src/components/pages-shared/onboarding/IntegrationExplorer/components/QuickInstallDialog.tsx
